Core: fix Label line break mode in Gtk
Made label widget in Gtk respect LineBreakMode and MaxLines
properties, the way it does on other platforms. By doing this
labels won't crop text with dashes in any part of the word and
they keep the words complete.
These properties were added to ILabel interface, as it is in
MAUI main branch.