GithubCI: build nuget packages
* Generate & upload stable release package for tags.
* Upload pre-release versions for main branch commits,
using nugetPreRelease.fsx to get a version number that
includes the date and commit hash inside it.

These packages will only be valid for the GTK platform
so when building against nonLinux platforms the consumers
of Lima will need to switch to restore and link to MAUI
nuget packages (instead of LIMA) at compile time.